- Added comments in EFRender .cpp to make code legible -Added comments in AppWindow.cpp to make code legible. - Added selection code . - Linked selection code to XYZ Arrows. - Added Function to calculate translation based on mouse position for arrows xyz. - Tested all 3 axis and translations and calculations work. - Added in code for to parse Scale XYZ shape. - Tested scale xyz shape and it looks fine.

+++ EnviroFlow/EFCreationSystem.h
@@ -52,7 +52,7 @@ struct EF_MESH {
 	GLfloat xyz[3];
 	GLubyte rgba[4];
 	GLfloat rot[4];
-	GLfloat pos[3];
+	GLdouble pos[3];
 	GLfloat scale[3];
 
 	EF_QUAD *m_quad;
